clean up

+++ services/moveService.js
@@ -1,7 +1,12 @@
 const moveRepository = require("../repositories/moveRepository");
 const accountService = require("./accountService");
 const historyService = require("./historyService");
-const { MOVE_SUBTYPES, MOVE_TYPES } = require("../constants");
+const {
+  MOVE_SUBTYPES,
+  MOVE_TYPES,
+  PERIOD_VALUES,
+  ACCOUNT_TYPES,
+} = require("../constants");
 const {
   getTodayRange,
   getYesterdayRange,
@@ -16,19 +21,19 @@ class MoveService {
   async getMovesByPeriod(period, subType, shopId) {
     let query = { shopId };
 
-    if (period === "daily") {
+    if (period === PERIOD_VALUES.daily) {
       const { start, end } = getTodayRange();
       query.date = { $gte: start, $lte: end };
     }
-    if (period === "yesterday") {
+    if (period === PERIOD_VALUES.yesterday) {
       const { start, end } = getYesterdayRange();
       query.date = { $gte: start, $lt: end };
     }
-    if (period === "weekly") {
+    if (period === PERIOD_VALUES.weekly) {
       const { start, end } = getWeekRange();
       query.date = { $gte: start, $lt: end };
     }
-    if (period === "monthly") {
+    if (period === PERIOD_VALUES.monthly) {
       const { start, end } = getMonthRange();
       query.date = { $gte: start, $lt: end };
     }
@@ -86,9 +91,19 @@ class MoveService {
     if (!session) return;
 
     try {
-      await session.startTransaction();
-
       const { type, amount, account, description, subType, accountId } = data;
+
+      const accounts = await accountService.getAccounts(user.shopId);
+      const primaryAccount = accounts.find(
+        (acc) => acc.type === ACCOUNT_TYPES.primary,
+      );
+      const moveAccount = accounts.find((acc) => acc.name === account);
+
+      if (!primaryAccount || !moveAccount) {
+        throw error("error finding primaryAccount or moveAccount");
+      }
+
+      await session.startTransaction();
 
       const move = new Move({
         type,
@@ -105,11 +120,6 @@ class MoveService {
       });
 
       const createdMove = await moveRepository.save(move, session);
-
-      const accounts = await accountService.getAccounts(user.shopId, session);
-
-      const primaryAccount = accounts.find((acc) => acc.type === "primary");
-      const moveAccount = accounts.find((acc) => acc.name === account);
 
       if (subType === MOVE_SUBTYPES.win) {
         await accountService.updateAccount(
@@ -225,17 +235,27 @@ class MoveService {
     if (!session) return;
 
     try {
-      await session.startTransaction();
+      const move = await moveRepository.findById(moveId);
 
-      const move = await moveRepository.findById(moveId, session);
+      if (!move) {
+        throw error("Invalid move id");
+      }
+
       const { subType, amount, account, shop } = move;
 
-      await moveRepository.deleteById(moveId, session);
-
-      const accounts = await accountService.getAccounts(user.shopId, session);
-
-      const primaryAccount = accounts.find((acc) => acc.type === "primary");
+      const accounts = await accountService.getAccounts(user.shopId);
+      const primaryAccount = accounts.find(
+        (acc) => acc.type === ACCOUNT_TYPES.primary,
+      );
       const moveAccount = accounts.find((acc) => acc.name === account);
+
+      if (!primaryAccount || !moveAccount) {
+        throw error("error finding primaryAccount or moveAccount");
+      }
+
+      await session.startTransaction();
+
+      await moveRepository.deleteById(moveId, session);
 
       if (subType === MOVE_SUBTYPES.win) {
         await accountService.updateAccount(
